requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.8.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==3.3
iniconfig==1.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.2.1
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
